+++ app/src/main/java/com/example/plantoplant/navigation/CalenderFragment.kt
@@ -12,6 +12,19 @@ import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.plantoplant.CalenderAdapter
 import com.example.plantoplant.R
+import com.example.plantoplant.ToDoMonths
+import com.example.plantoplant.util.ServerCon
+import kotlinx.coroutines.*
+import org.json.JSONArray
+import org.json.JSONException
+import org.json.JSONObject
+import java.io.BufferedReader
+import java.io.IOException
+import java.io.InputStreamReader
+import java.io.OutputStreamWriter
+import java.net.HttpURLConnection
+import java.net.MalformedURLException
+import java.net.URL
 import java.time.LocalDate
 import java.time.YearMonth
 import java.time.format.DateTimeFormatter
@@ -56,10 +69,26 @@ class CalenderFragment : Fragment() {
         monthYearText.text = monthYearFromDate(selectedDate) //년월 텍스트뷰
         userId = arguments?.getString("email") ?: "" // 유저 아이디
 
+        val year = selectedDate.year
+        val month = selectedDate.month.value
+        val selectedDay = selectedDate.dayOfMonth
+        var toDoArray = ArrayList<ToDoMonths>()
+
+        val job = CoroutineScope(Dispatchers.IO).launch {
+            val response = async { (getCalendarTodo(userId, year, month, selectedDay)) }
+            val jsonResult = response.await()
+            toDoArray = parseJson(jsonResult)
+        }
+
+        runBlocking {
+            job.join()
+            job.cancel()
+        }
+
         //날짜 생성, 리스트 담기
         val dayList = dayInMonthArray(selectedDate)
         //어댑터 초기화
-        val adapter = CalenderAdapter(dayList, userId, selectedDate, requireContext())
+        val adapter = CalenderAdapter(dayList, toDoArray, requireContext())
         //레이아웃 설정(열 7개)
         val manager: RecyclerView.LayoutManager = GridLayoutManager(requireContext(), 7)
         //레이아웃 적용
@@ -78,7 +107,6 @@ class CalenderFragment : Fragment() {
     //날짜 생성
     private fun dayInMonthArray(date: LocalDate): ArrayList<String>{
         var dayList = ArrayList<String>()
-
         var yearMonth = YearMonth.from(date)
         //해당 월 마지막 날짜 가져오기
         var lastDay = yearMonth.lengthOfMonth()
@@ -98,4 +126,75 @@ class CalenderFragment : Fragment() {
         return dayList
     }
 
+    // 한 달치 할 일 목록 가져오기
+    private fun getCalendarTodo(userId: String, year: Int, month: Int, selectedDay: Int): String {
+        var response = ""
+
+        try {
+            val con = ServerCon()
+            val url = URL(con.url + "calendar/months")
+            val conn = url.openConnection() as HttpURLConnection
+            conn.defaultUseCaches = false
+            conn.doInput = true
+            conn.doOutput = true
+            conn.requestMethod = "POST"
+            conn.setRequestProperty("Content-Type", "application/json")
+            conn.setRequestProperty("Accept", "application/json")
+
+            val jsonObject = JSONObject()
+            jsonObject.put("year", year)
+            jsonObject.put("month", month)
+            jsonObject.put("day", selectedDay)
+            jsonObject.put("user_id", userId)
+
+            // 서버로 값 전송
+            val outStream = OutputStreamWriter(conn.outputStream, "UTF-8")
+            outStream.write(jsonObject.toString())
+            outStream.flush()
+
+            // 서버에서 결과 받기
+            val inputStream = conn.inputStream
+            val bufferedReader = BufferedReader(InputStreamReader(inputStream))
+            val stringBuilder = StringBuilder()
+            var line: String?
+            while (bufferedReader.readLine().also { line = it } != null) {
+                stringBuilder.append(line).append('\n')
+            }
+
+            response = stringBuilder.toString()
+
+            println(response)
+        } catch (e: MalformedURLException) {
+            e.printStackTrace()
+        } catch (e: IOException) {
+            e.printStackTrace()
+        }
+
+        return response
+    }
+
+    // json 파싱
+    private fun parseJson(json: String): ArrayList<ToDoMonths> {
+        var result = ArrayList<ToDoMonths>()
+
+        try {
+            val jsonArray = JSONArray(json)
+
+            for (i in 0 until jsonArray.length()) {
+                val tdm: JSONObject = jsonArray.getJSONObject(i)
+                val toDoMonths = ToDoMonths(
+                    tdm.getInt("id"),
+                    LocalDate.parse(tdm.getString("date"), DateTimeFormatter.ISO_DATE),
+                    tdm.getString("toDo"),
+                    tdm.getBoolean("toDoCompleted")
+                )
+
+                result.add(toDoMonths)
+            }
+        } catch (e: JSONException) {
+            e.printStackTrace()
+        }
+
+        return result
+    }
 }
