+++ app/src/main/java/com/example/plantoplant/util/ServerCon.kt
@@ -1,5 +1,5 @@
 package com.example.plantoplant.util
 
 class ServerCon {
-    val url: String = "http://223.194.129.81:8080/"
+    val url: String = "http://192.168.0.18:8080/"
 }
